requirements.txt
certifi==2019.11.28
chardet==3.0.4
enum34==1.1.6
idna==2.8
Pillow==6.2.2
pynput==1.6.5
python-xlib==0.26
requests==2.22.0
selenium==3.141.0
six==1.14.0
urllib3==1.25.7
